update usage file
