--- lib/core/lib/generators/generators.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require_relative './specs_generator.rb'
-
-module Generators
-  include SpecsGenerator
-end
