--- lib/core/lib/generators/request_specs/request_specs_generator.rb
@@ -1,9 +1,0 @@
-require_relative '../generators.rb'
-
-class RequestSpecsGenerator < Rails::Generators::NamedBase
-  include Generators
-
-  def create_files
-    create_request_specs
-  end
-end
