+++ lib/core/lib/generators/ros/generators.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+# Dir['./**/*.rb'].each { |f| require f }
+
+require_relative './request_specs/request_specs_generator.rb'
